add real jobs data to redux (and refactor a little)
